MediaController와 Media 상태를 관리하는 PlaybackController 구현 및 UI 연동

+++ data/src/main/java/com/yongjin/musicplayer/data/MediaDataSource.kt
@@ -96,7 +96,7 @@ class MediaDataSource @Inject constructor(
                 val thumbnailUri = ContentUris.withAppendedId(EXTERNAL_ALBUM_ART_URI, albumId)
                 val contentUri = ContentUris.withAppendedId(uri, id)
                 songs.add(
-                    Song(id, album, artist, title, duration, track, thumbnailUri, contentUri)
+                    Song(id, album, artist, title, duration.toLong(), track, thumbnailUri, contentUri)
                 )
             }
         }
